requirements.txt
pystreamable
discord.py
python-dotenv
requests
twitch-dl
urlextract
spaw
